Small optimisations of `rd_to_date`.
As per "Section 8 / First column" of the paper, we can replace
    let n = n % 146097 / 4;
    let n = 4 * n + 3;
with
    let r = n % 146097;
    let n = r | 3;

which might improve the generated code. Unfortunately the Rust compiler
doesn't seem to be able to do it on its own.

We can replace
    let m = m - 12 * nd;
with
    let m = if nd == 1 { m - 12 } else { m };

Fortunately, for the first expression above the compiler does not emit
`imul` and it does not branch. Indeed, it uses a `cmovb` which is good.
However, the same happens for the expression using the ternary operator.
There's a difference though. Recall that
    let nd = (n >= 306) as u32;
and `nd` is used in the calculation of `y` and `m`. For some reason, for the
first expression, the compiler twice emits the `cmp` instruction to perform
the (equivalent of the) comparison with `306`. For the second expression,
only one `cmp` is emitted.