chore: More reliable benchmarks by array input
Previous benchmarking method picked a random value and repeatedly
tested the same value for each sample, leading to a case where
performance differences due to different input values were marked
as outliers.

Current method pregenerates a random array of values to be processed
so each sample will contain a mix of input values, leading to them
being averaged based on how often they appear in the randomized data.

This will cause an increase of 50 ps for overhead, which is not
relevant for our benchmarking.